Reorganize Edirol SD-90 MusE IDF output
The initial implementation made groups for the product of the 4 set
(Classical, Contemporary, Solo, Enhanced) and each of the 16 GM derived
instrument groups. 64 PatchGroups is too unwieldy in the UI, so this
alternative has been made. There are now the 16 subdivisions for the
instrument groupings, and all sets and variations are placed in that
one group. They are sorted by PC and then by set. This keeps the variations
from a single set close together.